+++ .gitlab-ci.yml
@@ -18,4 +18,4 @@ deploy:
     SERVER: gramener.com
     URL: comicgen
     VERSION: static
-    SETUP: yarn install && npm run build && bash png.sh
+    SETUP: yarn install && npm run build && node scripts/png.js && bash scripts/compress.sh
